backend/requirements.txt
fastapi==0.114.0
uvicorn[standard]==0.30.0
SQLAlchemy==2.0.30
psycopg[binary]==3.2.1
httpx==0.27.0
python-multipart==0.0.9
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
pydantic==2.8.2
authlib==1.3.1
pypdf==4.2.0
email-validator==2.1.1
youtube-transcript-api==1.0.3
python-pptx==1.0.2
